lib/httparty.rb: Improve HTTParty::Unsuported msg.
If an unsupported format is specified a HTTParty::Unsupported exception is
raised along with an error message enumerating the allowed formats. Valid
formats are specified as symbols, e.g. :yaml, but the error message
stringifies the symbols. This can be misleading.

For example, if a user specifies `format
'yaml'` he is confronted with an error message reading "Must be one of:
json, xml, html, yaml, plain". From his perspective, his format is in this
list, so the error is confusing.

This patch simply modifies the error message to read "Must be one of: :json,
:xml, :html, :yaml, :plain", which is hopefully clearer.