Documented interaction between .get and base_uri.
If a user specifies a base_uri with a trailing slash, he may think he can pass
.get a path without the leading slash under the belief that the path is simply
appended to the base URI. In reality, when the base_uri is normalised the
trailing slash is removed, breaking this assumption.

I have expanded the Rdoc for the .get method to explain this gotcha, although
we should probably do more. It may make sense to always concatenate the
base_uri and path unless the path starts with a protocol specification.